fix(#1961): cast nullable instrument_id filter param in get_activity (#1962)
Portfolio → Activity tab 500'd on every load. #1926 (5616d68b) added an
optional instrument_id filter to both get_activity queries:

    AND (%(instrument_id)s IS NULL OR instrument_id = %(instrument_id)s)

The per-instrument Positions tab supplies a concrete int, so Postgres infers
the param type. The Portfolio-level Activity tab supplies none → the param
binds as an untyped NULL (OID 0); psycopg3's extended-protocol send gives the
planner no type context and it aborts the whole statement:

    psycopg.errors.AmbiguousParameter: could not determine data type of parameter $2

Fix: cast every occurrence to the column type (trade_events.instrument_id is
bigint): %(instrument_id)s::bigint. Applied to the count and page queries.

- Regression guard: tests/test_api_portfolio_activity_null_filter_db.py
  (db-tier — a real planner; the mocked-cursor tests in test_api_portfolio.py
  cannot reproduce a type-inference error). Seeds two instruments: no-filter
  returns both, instrument_id filters to one. Verified it fails with
  AmbiguousParameter against the un-cast code.
- Lesson extracted: sql-correctness skill §Parameterisation + review-prevention-log
  (nullable-filter param must be cast to its column type).

+++ app/api/portfolio.py
@@ -1241,7 +1241,7 @@ def get_activity(
             SELECT COUNT(*) AS total
             FROM trade_events
             WHERE (%(include_mirrors)s OR COALESCE(social_trade_id, 0) = 0)
-              AND (%(instrument_id)s IS NULL OR instrument_id = %(instrument_id)s)
+              AND (%(instrument_id)s::bigint IS NULL OR instrument_id = %(instrument_id)s::bigint)
             """,
             {"include_mirrors": include_mirrors, "instrument_id": instrument_id},
         ).fetchone()
@@ -1268,7 +1268,7 @@ def get_activity(
                 WHERE event_kind = 'open'
             ) o ON o.position_id = te.position_id AND te.event_kind = 'close'
             WHERE (%(include_mirrors)s OR COALESCE(te.social_trade_id, 0) = 0)
-              AND (%(instrument_id)s IS NULL OR te.instrument_id = %(instrument_id)s)
+              AND (%(instrument_id)s::bigint IS NULL OR te.instrument_id = %(instrument_id)s::bigint)
             ORDER BY te.executed_at DESC, te.event_id DESC
             LIMIT %(limit)s
             """,
